Don't raise TransportError if disconnect fails in Client.close
Even if the disconnect request fails due to network errors, the
server will eventually close the client side session after a
certain amount of time. So instead of re-raising the
TransportError, the close method will catch this error and log
the failure. The important thing is that the client did it's
best to notify the server.