Raise error in Client.receive when the connection is lost
The transport will recover from temporary network failures if
connection is lost with the server, which works transparently
from the client's point of view. The downside is that if the
server becomes unavailable for a longer amount of time then the
transport will try to reconnect indefinitely without the client
ever knowing about the problem. Define a connection_timeout
variable in Client, and when the receive method is called (the
client waits for messages) raise an error if the connection
drops out for more than the time specified by connection_timeout.