Handle connection timeouts properly with WebsocketTransport
A websocket becomes aware that it has been closed only when it
receives a CLOSE message. So websockets which are not used
continuously for receiving might only get closed after a the
transports tries to send new messages and it's waiting for a
response.
To solve these issues try to get a new open websocket in
WebsocketTransport._send_final_payload and resend the messages
if a TransportConnectionClosed error is raised.