Handle server side disconnects without the incoming message queue
Relying on the incoming message queue to forward responses about
server side disconnects from the transport to the client is
problematic since the queue has a limited size and these responses
in some cases wouldn't even reach the client.
To solve this issue, the client should wait for the
SERVER_DISCONNECTED state and raise an error from the result of
the last connect result.