Validate subscription error responses more rigorously
If a subscription request fails with an error response there is
a high probability that the name of the channel won't be included
in the response. Check the existance of the subscription key in
the response before trying to read it. If the key is not present
then simply ignore the response and don't alter the set of
subscriptions.