Automatically reconnect if no advice is provided in the response
Even though the default implementation of the CometD server
returns a reconnect advice field for every connect and handshake
response, stop relying on this behavior since according to the
specifications the advice field is optional.
Only check for the presence of the advice field and the reconnect
advice when a connect or handshake request fails, and if no advice
is provided try to reconnect automatically.